refactor: extract jwt logic
