+++ tests/test_jwt_tools.py
@@ -4,6 +4,7 @@ import json
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 import app
+from retrorecon import jwt_service
 
 
 def setup_tmp(monkeypatch, tmp_path):
@@ -36,7 +37,7 @@ def test_jwt_decode_encode_roundtrip(tmp_path, monkeypatch):
     setup_tmp(monkeypatch, tmp_path)
     with app.app.test_client() as client:
         client.post('/new_db', data={'db_name': 'rt'})
-        demo = app.jwt.encode({'sub': '1'}, 'secret', algorithm='HS256')
+        demo = jwt_service.jwt.encode({'sub': '1'}, 'secret', algorithm='HS256')
         resp = client.post('/tools/jwt_decode', data={'token': demo})
         assert resp.status_code == 200
         data = resp.get_json()
@@ -51,14 +52,14 @@ def test_jwt_encode_handles_header_input(tmp_path, monkeypatch):
     setup_tmp(monkeypatch, tmp_path)
     with app.app.test_client() as client:
         client.post('/new_db', data={'db_name': 'rt2'})
-        token = app.jwt.encode({'iss': 'me'}, 'secret', algorithm='HS256')
+        token = jwt_service.jwt.encode({'iss': 'me'}, 'secret', algorithm='HS256')
         resp = client.post('/tools/jwt_decode', data={'token': token})
         data = resp.get_json()
         text = json.dumps(data['header'], indent=2) + "\n" + json.dumps(data['payload'], indent=2)
         resp2 = client.post('/tools/jwt_encode', data={'payload': text, 'secret': 'secret'})
         assert resp2.status_code == 200
         new_token = resp2.get_data(as_text=True)
-        decoded = app.jwt.decode(new_token, 'secret', algorithms=['HS256'])
+        decoded = jwt_service.jwt.decode(new_token, 'secret', algorithms=['HS256'])
         assert decoded['iss'] == 'me'
 
 
@@ -66,7 +67,7 @@ def test_jwt_decode_warnings_and_exp(tmp_path, monkeypatch):
     setup_tmp(monkeypatch, tmp_path)
     with app.app.test_client() as client:
         client.post('/new_db', data={'db_name': 'warn'})
-        none_token = app.jwt.encode({'sub': '1'}, '', algorithm='none')
+        none_token = jwt_service.jwt.encode({'sub': '1'}, '', algorithm='none')
         resp = client.post('/tools/jwt_decode', data={'token': none_token})
         assert resp.status_code == 200
         data = resp.get_json()
@@ -74,7 +75,7 @@ def test_jwt_decode_warnings_and_exp(tmp_path, monkeypatch):
 
         import datetime
         exp = int((datetime.datetime.now(datetime.timezone.utc) - datetime.timedelta(seconds=1)).timestamp())
-        token = app.jwt.encode({'exp': exp}, 'secret', algorithm='HS256')
+        token = jwt_service.jwt.encode({'exp': exp}, 'secret', algorithm='HS256')
         resp = client.post('/tools/jwt_decode', data={'token': token})
         data = resp.get_json()
         assert 'exp_readable' in data
@@ -86,7 +87,7 @@ def test_jwt_cookie_logging(tmp_path, monkeypatch):
     setup_tmp(monkeypatch, tmp_path)
     with app.app.test_client() as client:
         client.post('/new_db', data={'db_name': 'jwtlog'})
-        token = app.jwt.encode({'iss': 'me'}, '', algorithm='none')
+        token = jwt_service.jwt.encode({'iss': 'me'}, '', algorithm='none')
         resp = client.post('/tools/jwt_decode', data={'token': token})
         assert resp.status_code == 200
         resp = client.get('/jwt_cookies')
@@ -99,7 +100,7 @@ def test_jwt_cookie_edit_delete_export(tmp_path, monkeypatch):
     setup_tmp(monkeypatch, tmp_path)
     with app.app.test_client() as client:
         client.post('/new_db', data={'db_name': 'jwt2'})
-        token = app.jwt.encode({'iss': 'you'}, '', algorithm='none')
+        token = jwt_service.jwt.encode({'iss': 'you'}, '', algorithm='none')
         client.post('/tools/jwt_decode', data={'token': token})
         data = client.get('/jwt_cookies').get_json()
         jid = data[0]['id']
@@ -116,7 +117,7 @@ def test_jwt_cookie_edit_delete_export(tmp_path, monkeypatch):
 def test_jwt_decode_requires_db(tmp_path, monkeypatch):
     setup_tmp(monkeypatch, tmp_path)
     with app.app.test_client() as client:
-        token = app.jwt.encode({'sub': 1}, '', algorithm='none')
+        token = jwt_service.jwt.encode({'sub': 1}, '', algorithm='none')
         resp = client.post('/tools/jwt_decode', data={'token': token})
         assert resp.status_code == 400
         data = resp.get_json()
